+++ sm2/tsa/tests/test_stattools.py
@@ -5,16 +5,14 @@ import warnings
 
 import pytest
 import numpy as np
-from numpy.testing import assert_almost_equal, assert_equal, assert_allclose
+from numpy.testing import assert_almost_equal, assert_equal
 import pandas as pd
 
-from sm2.tools.sm_exceptions import ColinearityWarning, MissingDataError
 from sm2.compat.numpy import recarray_select
-from sm2.tsa.stattools import (adfuller, acf, pacf_yw,
+from sm2.tsa.stattools import (pacf_yw,
                                pacf, grangercausalitytests,
-                               coint, acovf, kpss,
                                arma_order_select_ic)
-from sm2.datasets import macrodata, sunspots
+from sm2.datasets import macrodata
 from sm2.tsa.arima_process import arma_generate_sample
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
@@ -30,116 +28,6 @@ DECIMAL_2 = 2
 
 
 @pytest.mark.not_vetted
-class CheckADF(object):
-    """
-    Test Augmented Dickey-Fuller
-
-    Test values taken from Stata.
-    """
-    levels = ['1%', '5%', '10%']
-    data = macrodata.load()
-    x = data.data['realgdp']
-    y = data.data['infl']
-
-    def test_teststat(self):
-        assert_almost_equal(self.res1[0],
-                            self.teststat,
-                            DECIMAL_5)
-
-    def test_pvalue(self):
-        assert_almost_equal(self.res1[1],
-                            self.pvalue,
-                            DECIMAL_5)
-
-    def test_critvalues(self):
-        critvalues = [self.res1[4][lev] for lev in self.levels]
-        assert_almost_equal(critvalues,
-                            self.critvalues,
-                            DECIMAL_2)
-
-
-# class TestADFConstantTrendSquared(CheckADF):
-#    pass
-# TODO: get test values from R?
-
-
-@pytest.mark.not_vetted
-class TestADFConstant(CheckADF):
-    """
-    Dickey-Fuller test for unit root
-    """
-    @classmethod
-    def setup_class(cls):
-        cls.res1 = adfuller(cls.x, regression="c", autolag=None, maxlag=4)
-        cls.teststat = .97505319
-        cls.pvalue = .99399563
-        cls.critvalues = [-3.476, -2.883, -2.573]
-
-
-@pytest.mark.not_vetted
-class TestADFConstantTrend(CheckADF):
-    @classmethod
-    def setup_class(cls):
-        cls.res1 = adfuller(cls.x, regression="ct", autolag=None, maxlag=4)
-        cls.teststat = -1.8566374
-        cls.pvalue = .67682968
-        cls.critvalues = [-4.007, -3.437, -3.137]
-
-
-@pytest.mark.not_vetted
-class TestADFNoConstant(CheckADF):
-    @classmethod
-    def setup_class(cls):
-        cls.res1 = adfuller(cls.x, regression="nc", autolag=None, maxlag=4)
-        cls.teststat = 3.5227498
-        cls.pvalue = .99999
-        # Stata does not return a p-value for noconstant.
-        # Tau^max in MacKinnon (1994) is missing, so it is
-        # assumed that its right-tail is well-behaved
-        cls.critvalues = [-2.587, -1.950, -1.617]
-
-
-# No Unit Root
-
-
-@pytest.mark.not_vetted
-class TestADFConstant2(CheckADF):
-    @classmethod
-    def setup_class(cls):
-        cls.res1 = adfuller(cls.y, regression="c", autolag=None, maxlag=1)
-        cls.teststat = -4.3346988
-        cls.pvalue = .00038661
-        cls.critvalues = [-3.476, -2.883, -2.573]
-
-
-@pytest.mark.not_vetted
-class TestADFConstantTrend2(CheckADF):
-    @classmethod
-    def setup_class(cls):
-        cls.res1 = adfuller(cls.y, regression="ct", autolag=None, maxlag=1)
-        cls.teststat = -4.425093
-        cls.pvalue = .00199633
-        cls.critvalues = [-4.006, -3.437, -3.137]
-
-
-@pytest.mark.not_vetted
-class TestADFNoConstant2(CheckADF):
-    @classmethod
-    def setup_class(cls):
-        cls.res1 = adfuller(cls.y, regression="nc", autolag=None, maxlag=1)
-        cls.teststat = -2.4511596
-        cls.pvalue = 0.013747
-        # Stata does not return a p-value for noconstant
-        # this value is just taken from our results
-        cls.critvalues = [-2.587, -1.950, -1.617]
-        _, _1, _2, cls.store = adfuller(cls.y, regression="nc", autolag=None,
-                                        maxlag=1, store=True)
-
-    def test_store_str(self):
-        assert self.store.__str__() == 'Augmented Dickey-Fuller Test Results'
-
-
-@pytest.mark.not_vetted
 class CheckCorrGram(object):
     """
     Set up for ACF, PACF tests.
@@ -149,99 +37,7 @@ class CheckCorrGram(object):
     results = results_corrgram
 
 
-@pytest.mark.not_vetted
-class TestACF(CheckCorrGram):
-    """
-    Test Autocorrelation Function
-    """
-    @classmethod
-    def setup_class(cls):
-        cls.acf = cls.results['acvar']
-        cls.qstat = cls.results['Q1']
-        cls.res1 = acf(cls.x, nlags=40, qstat=True, alpha=.05)
-        cls.confint_res = cls.results[['acvar_lb', 'acvar_ub']].as_matrix()
-
-    def test_acf(self):
-        assert_almost_equal(self.res1[0][1:41], self.acf, DECIMAL_8)
-
-    def test_confint(self):
-        centered = self.res1[1] - self.res1[1].mean(1)[:, None]
-        assert_almost_equal(centered[1:41], self.confint_res, DECIMAL_8)
-
-    def test_qstat(self):
-        assert_almost_equal(self.res1[2][:40], self.qstat, DECIMAL_3)
-        # 3 decimal places because of stata rounding
-
-    #def pvalue(self):
-    #     pass
-    # NOTE: shouldn't need testing if Q stat is correct
-
-
-@pytest.mark.not_vetted
-class TestACF_FFT(CheckCorrGram):
-    # Test Autocorrelation Function using FFT
-    @classmethod
-    def setup_class(cls):
-        cls.acf = cls.results['acvarfft']
-        cls.qstat = cls.results['Q1']
-        cls.res1 = acf(cls.x, nlags=40, qstat=True, fft=True)
-
-    def test_acf(self):
-        assert_almost_equal(self.res1[0][1:], self.acf, DECIMAL_8)
-
-    def test_qstat(self):
-        # TODO: why is res1/qstat 1 short
-        assert_almost_equal(self.res1[2], self.qstat, DECIMAL_3)
-
-
-class TestACFMissing(CheckCorrGram):
-    # Test Autocorrelation Function using Missing
-    @classmethod
-    def setup_class(cls):
-        cls.x = np.concatenate((np.array([np.nan]), cls.x))
-        cls.acf = cls.results['acvar']  # drop and conservative
-        cls.qstat = cls.results['Q1']
-        cls.res_drop = acf(cls.x, nlags=40, qstat=True, alpha=.05,
-                           missing='drop')
-        cls.res_conservative = acf(cls.x, nlags=40, qstat=True, alpha=.05,
-                                   missing='conservative')
-        cls.acf_none = np.empty(40) * np.nan  # lags 1 to 40 inclusive
-        cls.qstat_none = np.empty(40) * np.nan
-        cls.res_none = acf(cls.x, nlags=40, qstat=True, alpha=.05,
-                           missing='none')
-
-    def test_raise(self):
-        with pytest.raises(MissingDataError):
-            acf(self.x, nlags=40, qstat=True, alpha=0.5, missing='raise')
-
-    def test_acf_none(self):
-        assert_almost_equal(self.res_none[0][1:41],
-                            self.acf_none,
-                            DECIMAL_8)
-
-    def test_acf_drop(self):
-        assert_almost_equal(self.res_drop[0][1:41],
-                            self.acf,
-                            DECIMAL_8)
-
-    def test_acf_conservative(self):
-        assert_almost_equal(self.res_conservative[0][1:41],
-                            self.acf,
-                            DECIMAL_8)
-
-    def test_qstat_none(self):
-        # TODO: why is res1/qstat 1 short
-        assert_almost_equal(self.res_none[2],
-                            self.qstat_none,
-                            DECIMAL_3)
-
-    # TODO: how to do this test? the correct q_stat depends on
-    # whether nobs=len(x) is used when x contains NaNs or whether
-    # nobs<len(x) when x contains NaNs
-    #def test_qstat_drop(self):
-    #    assert_almost_equal(self.res_drop[2][:40], self.qstat, DECIMAL_3)
-
-
+# TODO: eventually belongs in test_autocov (once pacf is there)
 @pytest.mark.not_vetted
 class TestPACF(CheckCorrGram):
     @classmethod
@@ -275,219 +71,7 @@ class TestPACF(CheckCorrGram):
         assert_almost_equal(pacfyw, pacfld, DECIMAL_8)
 
 
-# TODO: this doesn't produce the old results anymore
-@pytest.mark.not_vetted
-class TestCoint_t(object):
-    """
-    Get AR(1) parameter on residuals
-
-    Test Cointegration Test Results for 2-variable system
-
-    Test values taken from Stata
-    """
-    levels = ['1%', '5%', '10%']
-    data = macrodata.load()
-    y1 = data.data['realcons']
-    y2 = data.data['realgdp']
-
-    @classmethod
-    def setup_class(cls):
-        cls.coint_t = coint(cls.y1, cls.y2, trend="c",
-                            maxlag=0, autolag=None)[0]
-        cls.teststat = -1.8208817
-        cls.teststat = -1.830170986148
-
-    def test_tstat(self):
-        assert_almost_equal(self.coint_t, self.teststat, DECIMAL_4)
-
-
-@pytest.mark.not_vetted
-class TestKPSS(object):
-    """
-    R-code
-    ------
-    library(tseries)
-    kpss.stat(x, "Level")
-    kpss.stat(x, "Trend")
-
-    In this context, x is the vector containing the
-    macrodata['realgdp'] series.
-    """
-    data = macrodata.load()
-    x = data.data['realgdp']
-
-    def test_fail_nonvector_input(self):
-        with warnings.catch_warnings(record=True):
-            kpss(self.x)  # should be fine
-
-        x = np.random.rand(20, 2)
-        with pytest.raises(ValueError):
-            kpss(x)
-
-    def test_fail_unclear_hypothesis(self):
-        # these should be fine,
-        with warnings.catch_warnings(record=True):
-            kpss(self.x, 'c')
-            kpss(self.x, 'C')
-            kpss(self.x, 'ct')
-            kpss(self.x, 'CT')
-
-        with pytest.raises(ValueError):
-            kpss(self.x, "unclear hypothesis")
-
-    def test_teststat(self):
-        with warnings.catch_warnings(record=True):
-            kpss_stat, pval, lags, crits = kpss(self.x, 'c', 3)
-        assert_almost_equal(kpss_stat, 5.0169, DECIMAL_3)
-
-        with warnings.catch_warnings(record=True):
-            kpss_stat, pval, lags, crits = kpss(self.x, 'ct', 3)
-        assert_almost_equal(kpss_stat, 1.1828, DECIMAL_3)
-
-    def test_pval(self):
-        with warnings.catch_warnings(record=True):
-            kpss_stat, pval, lags, crits = kpss(self.x, 'c', 3)
-        assert pval == 0.01
-
-        with warnings.catch_warnings(record=True):
-            kpss_stat, pval, lags, crits = kpss(self.x, 'ct', 3)
-        assert pval == 0.01
-
-    def test_store(self):
-        with warnings.catch_warnings(record=True):
-            kpss_stat, pval, crit, store = kpss(self.x, 'c', 3, True)
-
-        # assert attributes, and make sure they're correct
-        assert store.nobs == len(self.x)
-        assert store.lags == 3
-
-    def test_lags(self):
-        with warnings.catch_warnings(record=True):
-            kpss_stat, pval, lags, crits = kpss(self.x, 'c')
-
-        assert_equal(lags,
-                     int(np.ceil(12. * np.power(len(self.x) / 100., 1 / 4.))))
-        # assert_warns(UserWarning, kpss, self.x)
-
-
 # -----------------------------------------------------------------
-
-@pytest.mark.not_vetted
-def test_coint():
-    nobs = 200
-    scale_e = 1
-    const = [1, 0, 0.5, 0]
-    np.random.seed(123)
-    unit = np.random.randn(nobs).cumsum()
-    y = scale_e * np.random.randn(nobs, 4)
-    y[:, :2] += unit[:, None]
-    y += const
-    y = np.round(y, 4)
-
-    # results from Stata egranger
-    res_egranger = {}
-    # trend = 'ct'
-    res = res_egranger['ct'] = {}
-    res[0] = [-5.615251442239, -4.406102369132,
-              -3.82866685109, -3.532082997903]
-    res[1] = [-5.63591313706, -4.758609717199,
-              -4.179130554708, -3.880909696863]
-    res[2] = [-2.892029275027, -4.758609717199,
-              -4.179130554708, -3.880909696863]
-    res[3] = [-5.626932544079, -5.08363327039,
-              -4.502469783057, -4.2031051091]
-
-    # trend = 'c'
-    res = res_egranger['c'] = {}
-    # first critical value res[0][1] has a discrepancy starting at 4th decimal
-    res[0] = [-5.760696844656, -3.952043522638,
-              -3.367006313729, -3.065831247948]
-    # manually adjusted to have higher precision as in other cases
-    res[0][1] = -3.952321293401682
-    res[1] = [-5.781087068772, -4.367111915942,
-              -3.783961136005, -3.483501524709]
-    res[2] = [-2.477444137366, -4.367111915942,
-              -3.783961136005, -3.483501524709]
-    res[3] = [-5.778205811661, -4.735249216434,
-              -4.152738973763, -3.852480848968]
-
-    # trend = 'ctt'
-    res = res_egranger['ctt'] = {}
-    res[0] = [-5.644431269946, -4.796038299708,
-              -4.221469431008, -3.926472577178]
-    res[1] = [-5.665691609506, -5.111158174219,
-              -4.53317278104, -4.23601008516]
-    res[2] = [-3.161462374828, -5.111158174219,
-              -4.53317278104, -4.23601008516]
-    res[3] = [-5.657904558563, -5.406880189412,
-              -4.826111619543, -4.527090164875]
-
-    # The following for 'nc' are only regression test numbers
-    # trend = 'nc' not allowed in egranger
-    # trend = 'nc'
-    res = res_egranger['nc'] = {}
-    nan = np.nan  # shortcut for table
-    res[0] = [-3.7146175989071137, nan, nan, nan]
-    res[1] = [-3.8199323012888384, nan, nan, nan]
-    res[2] = [-1.6865000791270679, nan, nan, nan]
-    res[3] = [-3.7991270451873675, nan, nan, nan]
-
-    for trend in ['c', 'ct', 'ctt', 'nc']:
-        res1 = {}
-        res1[0] = coint(y[:, 0], y[:, 1],
-                        trend=trend, maxlag=4, autolag=None)
-        res1[1] = coint(y[:, 0], y[:, 1:3],
-                        trend=trend, maxlag=4, autolag=None)
-        res1[2] = coint(y[:, 0], y[:, 2:],
-                        trend=trend, maxlag=4, autolag=None)
-        res1[3] = coint(y[:, 0], y[:, 1:],
-                        trend=trend, maxlag=4, autolag=None)
-
-        for i in range(4):
-            res = res_egranger[trend]
-            assert_allclose(res1[i][0], res[i][0], rtol=1e-11)
-            assert_allclose(res1[i][2], res[i][1:], rtol=0, atol=6e-7)
-
-
-@pytest.mark.not_vetted
-def test_coint_identical_series():
-    nobs = 200
-    scale_e = 1
-    np.random.seed(123)
-    y = scale_e * np.random.randn(nobs)
-    warnings.simplefilter('always', ColinearityWarning)
-    with warnings.catch_warnings(record=True) as w:
-        c = coint(y, y, trend="c", maxlag=0, autolag=None)
-    assert len(w) == 1
-    assert c[0] == 0.0
-    # Limit of table
-    assert c[1] > .98
-
-
-@pytest.mark.not_vetted
-def test_acovf2d():
-    dta = sunspots.load_pandas().data
-    dta.index = pd.DatetimeIndex(start='1700', end='2009', freq='A')[:309]
-    del dta["YEAR"]
-    res = acovf(dta)
-    assert_equal(res, acovf(dta.values))
-
-    X = np.random.random((10, 2))
-    with pytest.raises(ValueError):
-        acovf(X)
-
-
-@pytest.mark.not_vetted
-def test_acovf_fft_vs_convolution():
-    np.random.seed(1)
-    q = np.random.normal(size=100)
-
-    for demean in [True, False]:
-        for unbiased in [True, False]:
-            F1 = acovf(q, demean=demean, unbiased=unbiased, fft=True)
-            F2 = acovf(q, demean=demean, unbiased=unbiased, fft=False)
-            assert_almost_equal(F1, F2, decimal=7)
-
 
 @pytest.mark.skip(reason="safe_arma_fit not ported from upstream")
 @pytest.mark.not_vetted
@@ -557,28 +141,6 @@ def test_arma_order_select_ic_failure():
 
 
 @pytest.mark.not_vetted
-def test_acf_fft_dataframe():
-    # GH#322
-    result = acf(sunspots.load_pandas().data[['SUNACTIVITY']], fft=True)[0]
-    assert result.ndim == 1
-
-
-@pytest.mark.not_vetted
-def test_coint_perfect_collinearity():
-    nobs = 200
-    scale_e = 1
-    np.random.seed(123)
-    x = scale_e * np.random.randn(nobs, 2)
-    y = 1 + x.sum(axis=1)
-    warnings.simplefilter('always', ColinearityWarning)
-    with warnings.catch_warnings(record=True):
-        c = coint(y, x, trend="c", maxlag=0, autolag=None)
-    assert c[0] == 0.0
-    # Limit of table
-    assert c[1] > .98
-
-
-@pytest.mark.not_vetted
 def test_grangercausality():
     # some example data
     mdata = macrodata.load().data
@@ -604,13 +166,3 @@ def test_granger_fails_on_nobs_check():
     grangercausalitytests(X, 2, verbose=False)  # This should pass.
     with pytest.raises(ValueError):
         grangercausalitytests(X, 3, verbose=False)
-
-
-def test_pandasacovf():
-    # test that passing Series vs ndarray to acovf doesn't affect results
-    # TODO: GH reference?
-    # TODO: Same test for other functions?
-    ser = pd.Series(list(range(1, 11)))
-    assert_allclose(acovf(ser),
-                    acovf(ser.values),
-                    rtol=1e-12)
